move classes to module
